requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
jinja2==3.1.4
requests==2.31.0
pandas==2.2.2
numpy==1.26.4
Pillow==10.4.0
